call-center-multi-agent-with-tools-langgraph/requirements.txt
langchain
langchain-openai
langgraph
langchain-community
pydantic
fastapi
uvicorn
streamlit
python-dotenv
requests